backend/requirements.txt
Flask>=3.1
opencv-python>=4.11
easyocr
pyserial>=3.5
numpy
requests>=2.32
